pyproject.toml
[project]
name = "rotom-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "python-telegram-bot[job-queue]>=22.5",
]
